add member and apply feedbacks

+++ src/app/components/account/account.component.ts
@@ -9,6 +9,7 @@ import { Label, SingleDataSet, monkeyPatchChartJsLegend, monkeyPatchChartJsToolt
 import { ChartType } from 'chart.js';
 import { TransactionComponent } from '../transaction/transaction.component';
 import { Router } from '@angular/router';
+import { environment } from '../../../environments/environment';
 
 @Component({
   selector: 'app-account',
@@ -16,6 +17,7 @@ import { Router } from '@angular/router';
   styleUrls: ['./account.component.css']
 })
 export class AccountComponent implements OnInit {
+  serverDomain = environment.serverDomain;
   // Define all the variable
   @Output() accounts: any
 
@@ -37,6 +39,7 @@ export class AccountComponent implements OnInit {
   labels = [];
   singleDataSetAf = [];
   labelsAf = [];
+  owners:any;
   public pieChartOptions: ChartOptions = {
     // responsive: true,
   };
@@ -47,6 +50,7 @@ export class AccountComponent implements OnInit {
   public pieChartData: SingleDataSet = this.singleDataSet;
   public pieChartType: ChartType = 'pie';
   public pieChartLegend = true;
+  // public pieChartPercentage = true;
   public pieChartPlugins = [];
 
 
@@ -73,6 +77,7 @@ export class AccountComponent implements OnInit {
 
   ngOnInit() {
     this.checkPerm();
+    this.loadMembers();
   }
 
   checkPerm() {
@@ -141,18 +146,47 @@ export class AccountComponent implements OnInit {
 
   // Delete Item From Server.
   delete(row) {
-    if (confirm('Are sure to delete?')) {
-      this.apiService.delete(row.id, 'acnt').subscribe(
-        result => {
-          this.deleteUI(row);
-          this.toast.show('Account deleted successfully!',
-            { classname: 'bg-warning text-light', delay: 5000 }	
-          );
+    this.apiService.loadAll('pay').subscribe(
+      result => {
+        let pays: any = result;
+        if (pays.filter(x => x.account == row.id).length > 0) {
+          this.toast.show("Delete not allowed for this Account!\nThis Account has assigned Payment.", { classname: 'bg-danger text-light', delay: 5000 });
         }
-      );
-    }
+        else{
+          if (confirm('Are sure to delete?')) {
+            this.apiService.delete(row.id, 'acnt').subscribe(
+              result => {
+                this.deleteUI(row);
+                this.toast.show('Account deleted successfully!',
+                  { classname: 'bg-warning text-light', delay: 5000 }	
+                );
+              }
+            );
+          }
+        }
+      }
+    )
+  }
+  loadMembers() {
+    this.apiService.loadAll('memb').subscribe(
+      (result: any) => {
+        this.owners = result;
+      },
+      error => {
+        console.log(error);
+      }
+    );
   }
 
+  findOwnerName(owner){
+    let o = this.owners.filter(x => x.id == owner)[0];
+    if (o) {
+      return o.first_name +' '+ o.last_name;
+    }
+    else{
+      return '';
+    }
+  }
   // Delete Item From UI
   deleteUI(row) {
     this.tableData = this.dataSource.data;
@@ -197,13 +231,25 @@ export class AddDialog {
   isFormValid: boolean = true;
   isNewFile: boolean = false;
   apiErr;
+  owners;
   constructor(
     private apiService: ApiService,
     private toast: ToastService,
     public dialogRef: MatDialogRef<any>,
     @Inject(MAT_DIALOG_DATA) public data: DialogData
-  ) { }
-
+  ) {
+    this.loadMembers();
+  }
+  loadMembers() {
+    this.apiService.loadAll('memb').subscribe(
+      (result: any) => {
+        this.owners = result;
+      },
+      error => {
+        console.log(error);
+      }
+    );
+  }
   accountFC = new FormGroup({
     label: new FormControl('', [Validators.required, Validators.minLength(5), Validators.maxLength(25)]),
     owner: new FormControl(''),
@@ -279,18 +325,30 @@ export class EditDialog implements OnInit{
   isNewFile: boolean = false;
 
   apiErr;
+  owners;
 
   constructor(
     private apiService: ApiService,
     public dialogRef: MatDialogRef<any>,
     @Inject(MAT_DIALOG_DATA) public data: DialogData
-  ) { }
-
+  ) { 
+    this.loadMembers();
+  }
+  loadMembers() {
+    this.apiService.loadAll('memb').subscribe(
+      (result: any) => {
+        this.owners = result;
+      },
+      error => {
+        console.log(error);
+      }
+    );
+  }
   accountFC = new FormGroup({
     label: new FormControl(''),
     owner: new FormControl(''),
-    balance: new FormControl(''),
-    currency: new FormControl(''),
+    balance: new FormControl({ value: '', disabled: true }),
+    currency: new FormControl({ value: '', disabled: true }),
     desc: new FormControl(''),
     status: new FormControl(''),
     file: new FormControl(Validators.required),
